Sceneflag N POIhandler data structure change
To adapt to the need for multiple scene flags for a button:
1. Sceneflag in POI class is now a list of string instead of string
2. ProjectPOI in POIHandler is now a list of string instead of a list
of list of string. Now POIs are not grouped by scenes in projectPOI
list. When generating buttons, we look at each POI in projectPOI and
loop through each POI’s scene flag list. This change greatly reduced
the complexity of saving and reading POIs.

In this commit, feature that merges existing buttons in editor is
finished