Started integrating pause functionality
Added a list to the POIMenuStateManager to hold references to any
scripts that need disabled while in the edit menu.
Also put a loop in the update to make sure the scripts in the list are
enabled and disabled at the correct time.  Also set the timescale to 0
if the edit menu is active.
Currently there is an issue when hitting Apply to Cancel to exit edit
mode where the framerate drops considerably.