Added Scene Flag List
After a button has been selected in Edit mode, the Scene Flags list will
be filled with all of the scenes that were included in the build
settings.

There is currently an issue with the vertical layout group where the
toggles are being placed on top of each other instead of spread out.