Started incorporating MiniMap into Lynn
Started getting the minimap moved over to the new UI system.  Trying to
update the code and make it cleaner/more efficient.

Zoom Butttons currently work along with clicking on the minimap to
teleport.

Still need to fix compass size and position, and buttons to change the
minimap size.

Also need to reimplement full-screen option, because that appeared to be
missing from the package.