MiniMap Upgrades
MiniMapManager script now holds the majority of the code for the
minimap.  There is a compassmanager script, but that will be
incorporated into MiniMapManager soon so that it is all one script.

Currently zoom and size buttons work, and the compass is positioned and
scaled to stay in the upper right corner of the minimap and stay the
same relative size.